+++ src/plugins/core/main.js
